Extract Stack from file
